FOP-1969: Support for unicode Surrogate pairs
git-svn-id: https://svn.apache.org/repos/asf/xmlgraphics/fop/trunk@1827168 13f79535-47bb-0310-9956-ffa450edef68

+++ fop-core/src/main/java/org/apache/fop/render/pcl/fonts/truetype/PCLTTFFontReader.java
@@ -627,7 +627,7 @@ public class PCLTTFFontReader extends PCLFontReader {
                     int nextOffset = 0;
                     int charCode = 0;
                     if (entry.getUnicodeIndex().size() > 0) {
-                        charCode = (Integer) entry.getUnicodeIndex().get(0);
+                        charCode = entry.getUnicodeIndex().get(0);
                     } else {
                         charCode = entry.getIndex();
                     }
@@ -743,7 +743,7 @@ public class PCLTTFFontReader extends PCLFontReader {
                     OFMtxEntry entry = mtx.get(i);
                     int charCode = 0;
                     if (entry.getUnicodeIndex().size() > 0) {
-                        charCode = (Integer) entry.getUnicodeIndex().get(0);
+                        charCode = entry.getUnicodeIndex().get(0);
                     } else {
                         charCode = entry.getIndex();
                     }
